Add Dockerfile to build container
This commit introduces a Dockerfile for building application
container. It installs minimal necessary Python interpreter
and pip3 for further dependencies installation. Also, because
of Flask requirement, the locale is installed and configured.
The required application files are added to /app directory
and the application dependencies are installed as well.
All unnecessary files are removed from the container.
Default command for container is set to launch the Flask app.